% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{expit}
\alias{expit}
\title{Take the expit of a number, the inverse of logit.}
\usage{
expit(x)
}
\arguments{
\item{x}{A number on the real line.}
}
\value{
The expit of \code{x}.
}
\description{
Take the expit of a number, the inverse of logit.
}
\examples{
expit(-5)
}
