% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLT.R
\name{predictLT}
\alias{predictLT}
\title{Predict a Life Table from Child or Child/Adult Mortality.}
\usage{
predictLT(sex, cm, smooth = TRUE, outlogit = FALSE, out5 = TRUE,
  am = NULL)
}
\arguments{
\item{sex}{Character: 'female' or 'male'.}

\item{cm}{Decimal: the input value(s) for 5q0; either a single value or a vector of values.}

\item{smooth}{Boolean: use either smooth or raw SVD-derived components. Default is TRUE.}

\item{outlogit}{Boolean: output either logit scale or natural scale 1qx values. Default is FALSE.}

\item{out5}{Boolean: if returning natural scale values and out5=TRUE, then return in five-year age groups, 5qx. Default is TRUE.}

\item{am}{Optional decimal: input value(s) for 45q15; either single value or vector of values.  If a vector, must have the same number of elements as cm.}
}
\value{
Data frame: predicted 1qx values for ages 0:109. Age 110 assumed to be 1.0 and not returned. Columns labeled with input child mortality values.
}
\description{
\code{svdComp5q0} predicts single-year, age-specific probabilities of dying (1qx) using the SVD-Comp mortality model indexed by child/child-adult mortality.
}
\examples{
predictLT("female",0.05)
predictLT("male",0.03,am=0.26)
predictLT("male",0.03,TRUE,TRUE,TRUE,0.26)
predictLT("male",c(0.03,0.01))
\dontrun{predictLT("male",c(0.03,0.01),am=0.3)}
}
\references{
\url{https://arxiv.org/abs/1612.01408}
}
\author{
Samuel J. Clark, \email{work@samclark.net}
}
