% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveYearQ.R
\name{q1to5}
\alias{q1to5}
\title{Convert Life Table Probabilities of Dying from 1-year to Standard 5-year Age Groups.}
\usage{
q1to5(q1)
}
\arguments{
\item{q1}{Decimal: the input values for 1qx; either a single vector or a matrix, age x life table.}
}
\value{
Data frame: equivalent values for 5qx.
}
\description{
\code{q1to5} converts 1-year life table probabilities of dying 1qx to standard 5-year age groups: 0, 1-4, 5-9, etc.
}
\examples{
q1 <- predictLT("female",0.08,out5=FALSE)
q1to5(q1)
}
\author{
Samuel J. Clark, \email{work@samclark.net}
}
