% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdComp5q0.R
\docType{package}
\name{svdComp5q0}
\alias{svdComp5q0}
\alias{svdComp5q0-package}
\title{svdComp5q0: Child or Child/Adult Mortality-indexed Model Life Tables.}
\description{
The svdComp5q0 package predicts full age schedules of mortality in single year or 5-year age groups from either child mortality (5q0) alone or both child and adult mortality (45q15).
Predictions are made using the SVD-Comp mortality model described in \url{https://arxiv.org/abs/1612.01408}.
}
\section{svdComp5q0 functions}{

\strong{predictLT} takes child or child/adult mortality and generates full age schedules of mortality.
}

