% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{logit}
\alias{logit}
\title{Calculate logit.}
\usage{
logit(x)
}
\arguments{
\item{x}{A number in the range (0,1).}
}
\value{
The logit of \code{x}.
}
\description{
Calculate logit.
}
\examples{
logit(0.5)
}
